+++ src/main/java/pl/plajer/buildbattle/handlers/sign/ArenaSign.java
@@ -21,6 +21,7 @@ package pl.plajer.buildbattle.handlers.sign;
 import java.lang.reflect.InvocationTargetException;
 
 import org.bukkit.Location;
+import org.bukkit.Material;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
 import org.bukkit.block.Sign;
@@ -51,10 +52,13 @@ public class ArenaSign {
   }
 
   private void setBehindBlock() {
-    if (plugin.is1_14_R1()) {
-      this.behind = getBlockBehind();
-    } else {
-      this.behind = getBlockBehindLegacy();
+    this.behind = null;
+    if (sign.getBlock().getType() == Material.WALL_SIGN) {
+      if (plugin.is1_14_R1()) {
+        this.behind = getBlockBehind();
+      } else {
+        this.behind = getBlockBehindLegacy();
+      }
     }
   }